implement #181
